+++ app/BackloggdMirror/Assets/changelog.md
@@ -6,6 +6,7 @@
 
 - [Emulation] Support for RetroArch, Dolphin, Cemu
 - [Emulation] Support for Game Boy, Game Boy Color, Game Boy Advance, DS, 3DS, NES, SNES, Nintendo 64, GameCube, Wii, Wii U, Mega Drive, Game Gear, Dreamcast, PlayStation, PS2, PS3, PSP, PS Vita, Xbox and Xbox 360
+- Apps and emulated games can be added to a blacklist, so they are no longer detected
 
 ### Fixed